# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/session": {
      "get": {
        "operationId": "AuthController_session",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Session info"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current session (from access token)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Logged in. Cookies set."
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Invalid credentials"
          }
        },
        "summary": "Login (sets access/refresh cookies; also returns 204)",
        "tags": [
          "Auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "204": {
            "description": "Refreshed. Cookies updated."
          },
          "401": {
            "description": "Missing/invalid refresh token"
          }
        },
        "summary": "Refresh session (rotates refresh token cookie)",
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/customers": {
      "post": {
        "operationId": "CustomersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer created"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "409": {
            "description": "Customer code already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create customer",
        "tags": [
          "Customers"
        ]
      },
      "get": {
        "operationId": "CustomersController_list",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Free-text search across multiple fields.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "Filter by name (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "required": false,
            "in": "query",
            "description": "Filter by code (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taxId",
            "required": false,
            "in": "query",
            "description": "Filter by taxId (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "email",
            "required": false,
            "in": "query",
            "description": "Filter by email (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "phone",
            "required": false,
            "in": "query",
            "description": "Filter by phone (contains, case-insensitive).",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Max items per page (1..100).",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Cursor for pagination (id < cursor).",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Optional. If provided, filters by active status.",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "IsActive",
            "required": false,
            "in": "query",
            "description": "Alias for isActive (prefer isActive).",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customers list"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List customers",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "CustomersController_getById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer detail"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get customer by id",
        "tags": [
          "Customers"
        ]
      },
      "patch": {
        "operationId": "CustomersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCustomerDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Customer updated"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Customer not found"
          },
          "409": {
            "description": "Customer code already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update customer",
        "tags": [
          "Customers"
        ]
      },
      "delete": {
        "operationId": "CustomersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Customer id (26 chars)",
            "schema": {
              "minLength": 26,
              "maxLength": 26,
              "example": "01J1QZQ0VQ8J7TQH0YV3A1BCDE",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Customer deactivated (soft delete). Returns deleted=true"
          },
          "400": {
            "description": "Validation error"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Customer not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate customer (soft delete)",
        "tags": [
          "Customers"
        ]
      }
    },
...
  },
  "info": {
    "title": "Stock SaaS API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Send a valid JWT access token. This API also accepts the access token via cookie (see cookie auth)."
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Optional. If present, the API will read the access token from this cookie."
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "tenantSlug": {
            "type": "string",
            "description": "Tenant slug.",
            "example": "acme"
          },
          "email": {
            "type": "string",
            "description": "User email.",
            "example": "admin@acme.com"
          },
          "password": {
            "type": "string",
            "description": "User password.",
            "example": "password123"
          }
        },
        "required": [
          "tenantSlug",
          "email",
          "password"
        ]
      },
...
      "CreateCustomerDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "Optional customer code. Must be unique per tenant when provided.",
            "example": "C001",
            "maxLength": 64
          },
          "type": {
            "type": "string",
            "description": "Customer type. Defaults to RETAIL.",
            "enum": [
              "RETAIL",
              "WHOLESALE"
            ],
            "default": "RETAIL"
          },
          "taxId": {
            "type": "string",
            "description": "Tax id / document number (future billing fields).",
            "example": "20123456789",
            "maxLength": 32
          },
          "taxIdType": {
            "type": "string",
            "description": "Tax id type (future billing fields).",
            "enum": [
              "CUIT",
              "CUIL",
              "DNI",
              "PASSPORT",
              "FOREIGN"
            ]
          },
          "vatCondition": {
            "type": "string",
            "description": "VAT condition (future billing fields).",
            "enum": [
              "REGISTERED",
              "MONOTAX",
              "EXEMPT",
              "FINAL_CONSUMER",
              "FOREIGN"
            ]
          },
          "name": {
            "type": "string",
            "description": "Customer display name.",
            "example": "Juan Perez",
            "minLength": 1,
            "maxLength": 200
          },
          "email": {
            "type": "string",
            "description": "Email address.",
            "example": "juan@example.com",
            "maxLength": 320
          },
          "phone": {
            "type": "string",
            "description": "Phone number.",
            "example": "+54 11 5555-5555",
            "maxLength": 32
          },
          "address": {
            "type": "string",
            "description": "Customer address (free text).",
            "example": "Av. Siempre Viva 742, CABA",
            "maxLength": 300
          },
          "notes": {
            "type": "string",
            "description": "Free text notes / observations.",
            "example": "Prefers WhatsApp for contact.",
            "maxLength": 5000
          },
          "isActive": {
            "type": "boolean",
            "description": "Soft active flag. Defaults to true.",
            "default": true
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCustomerDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "object",
            "description": "Customer code. Send null to clear it.",
            "example": "C001",
            "maxLength": 64,
            "nullable": true
          },
          "type": {
            "type": "string",
            "description": "Customer type.",
            "enum": [
              "RETAIL",
              "WHOLESALE"
            ]
          },
          "taxId": {
            "type": "object",
            "description": "Tax id / document number. Send null to clear it.",
            "example": "20123456789",
            "maxLength": 32,
            "nullable": true
          },
          "taxIdType": {
            "type": "string",
            "description": "Tax id type. Send null to clear it.",
            "enum": [
              "CUIT",
              "CUIL",
              "DNI",
              "PASSPORT",
              "FOREIGN"
            ],
            "nullable": true
          },
          "vatCondition": {
            "type": "string",
            "description": "VAT condition. Send null to clear it.",
            "enum": [
              "REGISTERED",
              "MONOTAX",
              "EXEMPT",
              "FINAL_CONSUMER",
              "FOREIGN"
            ],
            "nullable": true
          },
          "name": {
            "type": "string",
            "description": "Customer display name.",
            "example": "Juan Perez",
            "maxLength": 200
          },
          "email": {
            "type": "object",
            "description": "Email address. Send null to clear it.",
            "example": "juan@example.com",
            "maxLength": 320,
            "nullable": true
          },
          "phone": {
            "type": "object",
            "description": "Phone number. Send null to clear it.",
            "example": "+54 11 5555-5555",
            "maxLength": 32,
            "nullable": true
          },
          "address": {
            "type": "object",
            "description": "Customer address. Send null to clear it.",
            "example": "Av. Siempre Viva 742, CABA",
            "maxLength": 300,
            "nullable": true
          },
          "notes": {
            "type": "object",
            "description": "Free text notes. Send null to clear it.",
            "example": "Prefers WhatsApp for contact.",
            "maxLength": 5000,
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Soft active flag (can be used to reactivate)."
          }
        }
      },
...
    }
  }
}